Added is_even() function
Pretty self explanitory. This will be important later to make sure that
every round of the bracket has an even number of entries.
NOTE: maybe randomly choose where to place bye week if need to balance
round so that lowest element on list doesnt have an overly easy time of
getting to the final round?